### 2023-06-10 #5531 @dfm
I also want to note that this submission is made in parallel with a AAS publication. We need to remember to make reviewers aware of our [JOSS collaboration with AAS publishing](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing), and that as part of this collaboration, AAS publishing makes a small donation to the running costs of JOSS.
