### 2023-06-10 #5531 @tobin-wainer
> **Submitting author:** @tobin-wainer ([Tobin Wainer](http://orcid.org/0000-0001-6320-2230)) **Repository:** https://github.com/tobin-wainer/elk **Branch with paper.md** (empty if default branch): joss **Version:** v0.9 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Dan Foreman-Mackey
>
> ## Status
> [![status](https://camo.githubusercontent.com/ddf1951c07d7f76a61b5b37ae771a1fcd5fdeb09f5ae1033f280194f5c444d9c/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f37643534656333306365643438353561383035633430666134303561646630342f7374617475732e737667)](https://joss.theoj.org/papers/7d54ec30ced4855a805c40fa405adf04)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS @tobin-wainer. **Currently, there isn't a JOSS editor assigned** to your paper.
>
> @tobin-wainer if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot @editorialbot is here to help you find and assign reviewers and start the main review. To find out what @editorialbot can do for you type:
>
> ```
> @editorialbot commands
> ```

Potential Suggested Reviewers:
- Timothy Bkacke Miller; tbmiller-astro
- afeinstein20
- sgeorge91
- zspace
- jessie-dotson
- bliir
