+++ stripped_bib.bib
@@ -1,0 +1,306 @@
+@article{stetson_center_1994,
+	title = {The center of the core-cusp globular cluster {M15}: {CFHT} and {HST} {Observations}, {ALLFRAME} reductions},
+	volume = {106},
+	issn = {0004-6280},
+	shorttitle = {The center of the core-cusp globular cluster {M15}},
+	url = {http://adsabs.harvard.edu/abs/1994PASP..106..250S},
+	doi = {10.1086/133378},
+	abstract = {The central brightness cusps seen in some globular clusters are thought 
+to be the relics of a gravothermal core collapse that occurred sometime
+in the clusters' past. Recent observations show that the centers of such
+clusters are bluer than their outskirts, indicating that the stellar
+populations there are somehow different than those farther out,
+presumably as a result of unusual physical processes that took place in
+these extremely dense regions. Here I analyze a large body of digital
+imagery from the Canada-France-Hawaii Telescope and the Hubble Space
+Telescope to obtain color-magnitude and color-color diagrams for stars
+in the central two arcminutes of the prototypical core-cusp globular
+cluster M15 = NGC 7078 = C 2127 + 119. These data were reduced with a
+new computer program, named ALLFRAME, that is described in detail here
+for the first time. ALLFRAME makes simultaneous use of the geometric and
+photometric information from all images of a given field to derive a
+self-consistent set of positions and magnitudes for all detected
+starlike objects in that area of sky, thereby extending the range of
+magnitude and crowding conditions for which useful photometry is
+obtainable. I tentatively conclude that the color gradient in M15 is due
+to three distinct effects: (1) there is a deficiency of the brightest
+red giants in the central regions of the cluster; (2) the giant branch
+of the dominant cluster population shifts systematically toward the blue
+as the center of the cluster is approached; and (3) the very center of
+the cluster (radius approximately equal or less than 12 seconds)
+contains a large population of blue and yellow stragglers that occupy
+the area of the color-magnitude diagram between the main-sequence
+turnoff and the horizontal branch and between the extended blue
+horizontal branch and the subgiant branch; many of these appear to have
+a significant ultraviolet excess.},
+	urldate = {2021-03-31},
+	journal = {Publications of the Astronomical Society of the Pacific},
+	author = {Stetson, Peter B.},
+	month = mar,
+	year = {1994},
+	keywords = {Astronomical Photometry, Globular Clusters, Calibrating, Color-Color Diagram, Color-Magnitude Diagram, Data Reduction, Hubble Space Telescope},
+	pages = {250--280},
+	file = {Full Text PDF:/Users/Tobin/Zotero/storage/ZJ8YLR8C/Stetson - 1994 - The center of the core-cusp globular cluster M15 .pdf:application/pdf},
+}
+
+@article{clement_variable_2001,
+	title = {Variable {Stars} in {Galactic} {Globular} {Clusters}},
+	volume = {122},
+	issn = {0004-6256},
+	url = {http://adsabs.harvard.edu/abs/2001AJ....122.2587C},
+	doi = {10.1086/323719},
[…]
