### 2023-06-10 #5531 @tobin-wainer
Hi @dfm, Thanks for reaching out. I've updated the .bib files like you've asked. I've also read through the linked page about the joint submission. Is there anything else I need to do specifically for the joint submission? Thanks!
