### 2023-06-18 #5531 @arfon
@tobin-wainer – any chance you could limit the BibTeX entries to those you're _actually_ citing in your paper?
