### 2023-06-18 #5531 @arfon
@warrickball – this paper is being jointly published with AAS. As such, could you please make sure you follow these additional instructions: https://joss.readthedocs.io/en/latest/editing.html#handling-of-papers-published-together-with-aas-publishing
