### 2023-06-18 #5531 @TomWagg
Hey @arfon, @tobin-wainer is dealing with some personal issues at the moment so may be a little slow to reply, I'm second author and happy to answer questions in the meantime.

On the bibtex entries Tobin has already updated those after Dan asked I believe. Also, in case it is useful, [I wrote up a little script to do it](https://gist.github.com/TomWagg/d3c6abf4a3663c8f0770a106063240d9), you'd be welcome to integrate it with the JOSS workflow to do this automatically if you like.
