### 2023-06-30 #5605 @editorialbot
[REVIEW]: ELK: A python package for correcting, analyzing, and diagnosing TESS integrated light curves
**Submitting author:** @tobin-wainer (<a href="http://orcid.org/0000-0001-6320-2230">Tobin Wainer</a>)
**Repository:** https://github.com/tobin-wainer/elk
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0
