### 2023-07-11 #5605 @ryanoelkers
@tobin-wainer very nice work! thanks for putting together this package. here are a few comments on the open source paper.

1. It may be useful to define what "integrated" means in the context of light curves as opposed to a single source. Integrated has many definitions in astronomy.
2. line 56 and line 66 - TESS has observed hundreds of millions of stars, not just 200K. The FFIs are still TESS observations and were part of the primary mission. The 200K mentioned were the short-cadence targets which had their light curves reduced by NASA. You may want to be clear in that distinction. All of the short cadence targets are blended as well. The majority are bright so the contamination from other sources is low, but that doesn't mean there aren't highly blended sources in those 200K. The TIC did not prioritize stars for short cadence based on blending. The TIC prioritization was a combination of brightness, contamination, planet-to-star radii, and "special" considerations.
3. line 85 - "Then gives users..." I think this is a typo.
4. There are a few references which appear to have issues compiling in the GitHub version.
