### 2023-07-12 b5f0358
Update paper.md
Defining Integrated and how it applies
