+++ paper.md
@@ -38,6 +38,8 @@ aas-journal: Astrophysical Journal
 
 `elk` (int**E**grated **L**ight **K**urve) is an open-source Python package designed for downloading and correcting *integrated* light curves from the Full Frame Images (FFI) of TESS data. We provide novel approaches for the analysis of heavily blended light curves. Our correction is in the form of aperture photometry, building upon the methodology of the Python package `lightkurve` [@lightkurve_collaboration_lightkurve_2018]. 
 
+The term `integrated' refers to the method in which we are extracting blended photometry from an aperture, a practice common in the literature. Our approach does this extraction at many time steps to generate light curves, and we evaluate how the integrated light changes through time. Each pixel light curve is dertermined independantly, and then summed to create the integrated light curve.
+
 Our correction techniques are able to account for spatially-transient scattered light from the Earth and Moon (using singular value decomposition methods and principle component analysis), known TESS telescope systems (encoding in co-trending basis vectors) and long-period astrophysical variability (through the application of a basis-spline model). Each of these techniques are flexible within `elk` and can be tuned or turned off entirely.
 
 `elk` additionally has a range of features to analyze integrated light curves. We include a wide variety of variability metrics from across astronomical subfields, including the Stetson J statistic and von Neumann ratio [@stetson_center_1994; @von_neumann_distribution_1941]. `elk` features simple and useful diagnostic capabilities, with flexible functions for the creation and visualisation of bootstrapped Lomb-Scargle periodograms and autocorrelation functions. Moreover, one can use `elk` to locate the precise sky position of a variable signature causing a feature in an integrated light curve, and identify likely candidate stars from SIMBAD [@wenger_simbad_2000]. Though we designed `elk` with the intention of studying cluster variability, it has the flexibility to be applied to *any* integrated study.
