### 2023-07-12 3ee5118
Update paper.md
Fixed typo in line 85
