### 2023-07-12 da3ab4b
Update paper.bib
Adding missing references
