+++ paper.bib
@@ -43,6 +43,61 @@ a significant ultraviolet excess.},
 	file = {Full Text PDF:/Users/Tobin/Zotero/storage/ZJ8YLR8C/Stetson - 1994 - The center of the core-cusp globular cluster M15 .pdf:application/pdf},
 }
 
+@article{alonso-garcia_variable_2021,
+	title = {Variable stars in the {VVV} globular clusters. {II}. {NGC} 6441, {NGC} 6569, {NGC} 6626 ({M} 28), {NGC} 6656 ({M} 22), {2MASS}-{GC} 02, and {Terzan} 10},
+	volume = {651},
+	issn = {0004-6361},
+	url = {https://ui.adsabs.harvard.edu/abs/2021A&A...651A..47A/abstract},
+	doi = {10.1051/0004-6361/202140546},
+	abstract = {Context. The Galactic globular clusters (GGCs) located in the inner regions of the Milky Way suffer from high extinction that makes their observation challenging. High densities of field stars in their surroundings complicate their study even more. The VISTA Variables in the Via Lactea (VVV) survey provides a way to explore these GGCs in the near-infrared where extinction effects are highly diminished. {\textless}BR /{\textgreater} Aims: We conduct a search for variable stars in several inner GGCs, taking advantage of the unique multi-epoch, wide-field, near-infrared photometry provided by the VVV survey. We are especially interested in detecting classical pulsators that will help us constrain the physical parameters of these GGCs. In this paper, the second of a series, we focus on NGC 6656 (M 22), NGC 6626 (M 28), NGC 6569, and NGC 6441; these four massive GGCs have known variable sources, but quite different metallicities. We also revisit 2MASS-GC 02 and Terzan 10, the two GGCs studied in the first paper of this series. {\textless}BR /{\textgreater} Methods: We present an improved method and a new parameter that efficiently identify variable candidates in the GGCs. We also use the proper motions of those detected variable candidates and their positions in the sky and in the color-magnitude diagrams to assign membership to the GGCs. {\textless}BR /{\textgreater} Results: We identify and parametrize in the near-infrared numerous variable sources in the studied GGCs, cataloging tens of previously undetected variable stars. We recover many known classical pulsators in these clusters, including the vast majority of their fundamental mode RR Lyrae. We use these pulsators to obtain distances and extinctions toward these objects. Recalibrated period-luminosity-metallicity relations for the RR Lyrae bring the distances to these GGCs to a closer agreement with those reported by Gaia, except for NGC 6441, which is an uncommon Oosterhoff III GGC. Recovered proper motions for these GGCs also agree with those reported by Gaia, except for 2MASS-GC 02, the most reddened GGC in our sample, where the VVV near-infrared measurements provide a more accurate determination of its proper motions. {\textless}P /{\textgreater}Lightcurves and full Tables 2, 5, 6, 7, 8, and 9 are only available at the CDS via anonymous ftp to {\textless}A href="http://cdsarc.u-strasbg.fr/"{\textgreater}cdsarc.u-strasbg.fr{\textless}/A{\textgreater} (ftp://130.79.128.5) or via {\textless}A href="http://cdsarc.u-strasbg.fr/viz-bin/cat/J/A+A/651/A47"{\textgreater}http://cdsarc.u-strasbg.fr/viz-bin/cat/J/A+A/651/A47{\textless}/A{\textgreater}},
+	language = {en},
+	urldate = {2022-04-26},
+	journal = {Astronomy and Astrophysics},
[…]
