### 2023-07-12 baf3d91
Update paper.md
Reworking statements about the TIC, and TESS source selection
