+++ paper.md
@@ -52,9 +52,7 @@ One important property of stars is their intrinsic photometric variability, incl
 
 For environments where individual stars can not be resolved, the current standard approach to population analysis is through integrated light methods. However, the variability of unresolved stellar populations, i.e., of their integrated light, remain extremely limited. One analysis of unresolved field populations in M87 [@conroy_ubiquitous_2015] characterized the density of long-period variable stars, as a prediction for the age of the population. The main issue with using integrated light for variability encountered in this study was the blending of sources. There have been some efforts to address this issue (in TESS, relevant for this current work) for the light curves of individual resolved stars (e.g., @oelkers_precision_2018; @nardiello_psf-based_2020; @higgins_localizing_2022). The work presented here in `elk`, sidesteps these issues by working directly with the integrated light curves.
 
-While the primary mission of TESS was the discovery of exoplanets, the unrivaled precision of its time series photometry led to many significant advances in stellar science. For example, TESS has been influential in the recent surge of new astroseismology results (e.g., @handberg_tess_2021). TESS has observed nearly 200,000 stars from the TESS Input Catalog (TIC; @stassun_tess_2018), selected for the purpose of detecting small transiting planets. While the primary targets in TESS were selected to avoid blended sources, the FFIs contain tens of millions of sources, of which many are blended. These blended sources still contain bountiful astrophysical information and there have been many studies focused on blended star extraction (e.g., @oelkers_precision_2018; @nardiello_psf-based_2020; @higgins_localizing_2022). For the FFIs, more than 20 million sources for which relative photometry with $1\%$ photometric precision can be obtained [@ricker_transiting_2015, @huang_photometry_2020, @kunimoto_quick-look_2021]. These data provide a rich data set which have been used across astronomy subfields, and thus, the treatment of these data have been the topic of recent studies [@hattori_unpopular_2022].
-
-The primary mission of TESS was observing nearly 200,000 unblended stars from the TESS Input Catalog (TIC; @stassun_tess_2018), selected for the purpose of detecting small transiting planets. However, due to the large TESS pixels, the FFIs also contain tens of millions of blended astrophysical sources, and there have been many studies focused on blended star extraction (e.g., @oelkers_precision_2018; @nardiello_psf-based_2020; @higgins_localizing_2022). These studies have resulted in more than 20 million blended FFI sources with relative photometry down to $1\%$ photometric precision  [@ricker_transiting_2015, @huang_photometry_2020, @kunimoto_quick-look_2021]. 
+While the primary mission of TESS was the discovery of exoplanets, the unrivaled precision of its time series photometry led to many significant advances in stellar science. For example, TESS has been influential in the recent surge of new astroseismology results (e.g., @handberg_tess_2021). The 200,000 primary targets in the TESS input catalog (TIC, @stassun_tess_2018) were selected to detect exoplanets, and are thus not heavily blended, due to the large TESS pixels the FFIs contain tens of millions of sources, of which many are blended. These blended sources still contain bountiful astrophysical information and there have been many studies focused on blended star extraction (e.g., @oelkers_precision_2018; @nardiello_psf-based_2020; @higgins_localizing_2022). For the FFIs, more than 20 million sources for which relative photometry with $1\%$ photometric precision can be obtained [@ricker_transiting_2015, @huang_photometry_2020, @kunimoto_quick-look_2021]. These data provide a rich data set which have been used across astronomy subfields, and thus, the treatment of these data have been the topic of recent studies [@hattori_unpopular_2022].
 
 `elk` is a python package designed to harness the bountiful information available in blended FFI photometry in the form of aperture based light curves. This package is the first of its kind for the correction, analysis and diagnosis of *integrated* light curves.
 
