### 2023-07-12 a503d7c
Update paper.md
Fix typo
