### 2023-07-12 #5605 @tobin-wainer
Hi @ryanoelkers,

Thanks so much for your comments. I have addressed each of your comments by doing the following:

1. This is very true. We have have quite the time trying to decide how to define this. I have added a paragraph defining what integrated means in our study.

2. I have re-worked these paragraphs into one which clarifies this point.

3. This was a typo and it has been fixed.

4. When we trimmed down the .bib file it seems we accidentally removed some references that we actually needed. These have been fixed.
