### 2023-07-21 #5605 @ryanoelkers
thanks for addressing those points @tobin-wainer! the new draft of the open source paper looks good.
