### 2023-07-28 #5605 @ryanoelkers
@tobin-wainer i've looked at some of the documentation and tutorials (mainly the calculating lightcurve statistics tutorial). it looks nice so far! just two small issues i found (which may not be issues, but worth checking).

1. in the line "%config InlineBackend.figure_format = "retina"" throws an invalid syntax line for my PyCharm IDE.
2. the lc.stats["n_peaks"] line reports 2 for me, but is 3 in the tutorial. The other numbers match within floating point errors (so it is likely not a problem, just something you should be aware of)
