### 2023-08-01 a91a696
Delete un-needed ref from cosmic
