+++ docs/tutorials/refs.bib
@@ -1,246 +1,0 @@
-@ARTICLE{Robson+2019,
-       author = {{Robson}, Travis and {Cornish}, Neil J. and {Liu}, Chang},
-        title = "{The construction and use of LISA sensitivity curves}",
-      journal = {Classical and Quantum Gravity},
-     keywords = {Astrophysics - High Energy Astrophysical Phenomena, General Relativity and Quantum Cosmology},
-         year = 2019,
-        month = may,
-       volume = {36},
-       number = {10},
-          eid = {105011},
-        pages = {105011},
-          doi = {10.1088/1361-6382/ab1101},
-archivePrefix = {arXiv},
-       eprint = {1803.01944},
- primaryClass = {astro-ph.HE},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2019CQGra..36j5011R},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Finn&Thorne2000,
-       author = {{Finn}, Lee Samuel and {Thorne}, Kip S.},
-        title = "{Gravitational waves from a compact star in a circular, inspiral orbit, in the equatorial plane of a massive, spinning black hole, as observed by LISA}",
-      journal = {Physical Review D},
-     keywords = {04.30.Db, 04.80.Nn, 97.60.Lf, Wave generation and sources, Gravitational wave detectors and experiments, Black holes, General Relativity and Quantum Cosmology},
-         year = 2000,
-        month = dec,
-       volume = {62},
-       number = {12},
-          eid = {124021},
-        pages = {124021},
-          doi = {10.1103/PhysRevD.62.124021},
-archivePrefix = {arXiv},
-       eprint = {gr-qc/0007074},
- primaryClass = {gr-qc},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2000PhRvD..62l4021F},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Barack&Cutler2004,
-       author = {{Barack}, Leor and {Cutler}, Curt},
-        title = "{LISA capture sources: Approximate waveforms, signal-to-noise ratios, and parameter estimation accuracy}",
-      journal = {Physical Review D},
-     keywords = {04.80.Nn, 04.25.Nx, 04.30.Db, 04.80.Cc, Gravitational wave detectors and experiments, Post-Newtonian approximation, perturbation theory, related approximations, Wave generation and sources, Experimental tests of gravitational theories, General Relativity and Quantum Cosmology, Astrophysics},
-         year = 2004,
-        month = apr,
-       volume = {69},
-       number = {8},
-          eid = {082005},
-        pages = {082005},
-          doi = {10.1103/PhysRevD.69.082005},
-archivePrefix = {arXiv},
-       eprint = {gr-qc/0310125},
- primaryClass = {gr-qc},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2004PhRvD..69h2005B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Peters1963,
-       author = {{Peters}, P.~C. and {Mathews}, J.},
-        title = "{Gravitational Radiation from Point Masses in a Keplerian Orbit}",
-      journal = {Physical Review},
-         year = 1963,
-        month = jul,
-       volume = {131},
-       number = {1},
-        pages = {435-440},
-          doi = {10.1103/PhysRev.131.435},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/1963PhRv..131..435P},
[…]
