### 2023-08-01 4b07b1d
Update index.rst
Add link to ADS for the paper
