+++ docs/index.rst
@@ -16,7 +16,7 @@ elk
 
 Welcome to the int\ **E**\ grated **L**\ ight\ **K**\ urve Python package, or simply ``elk``!
 
-This package allows you to download, correct, analyze and visualize *integrated* light curves from TESS FFI data using the methods outlined in `Wainer+2023 <https://ui.adsabs.harvard.edu/>`_. While originally designed for analysis of star clusters, it can be used to create integrated light curves for any given aperture.
+This package allows you to download, correct, analyze and visualize *integrated* light curves from TESS FFI data using the methods outlined in `Wainer+2023 <https://ui.adsabs.harvard.edu/abs/2023arXiv230709510W/abstract>`_. While originally designed for analysis of star clusters, it can be used to create integrated light curves for any given aperture.
 
 Use the links below to learn how to install ``elk``, explore different features in our detailed tutorials or peruse our exhaustive API documentation. We're excited to see how you will apply ``elk`` to your own work!
 
