### 2023-08-16 2485957
Reverting numpy to the local version
