+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy
+numpy < 1.24
 pip
 matplotlib
 astropy
