### 2023-08-21 #5605 @warrickball
Hi @danhey, thanks again for agreeing to review.

As a reminder of the process, please read the "Reviewer instructions & questions" in the [first comment](https://github.com/openjournals/joss-reviews/issues/5263#issue-1627737752) above, and generate your checklists by commenting `@editorialbot generate my checklist` on this issue. As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

If you submit issues or pull requests on the software repository, please mention `openjournals/joss-reviews#4992` so that the issue/PR is linked to this thread. Please also feel free to comment and ask questions on this thread. JOSS editors have found it better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We usually aim for the review process to be completed within about 4-6 weeks but I know this one has been delayed by various conferences and holidays over the Northern Hemisphere's summer. JOSS reviews are iterative and the authors can start responding while you continue to review other parts of the submission.

Finally, don't hesitate to ask many any questions you might have about the process.
