### 2023-08-22 22947c2
Add AAS Paper citation
