+++ paper.bib
@@ -359,3 +359,17 @@ Type: article},
 	file = {Full Text PDF:/Users/Tobin/Zotero/storage/3NLKEDRJ/Wenger et al. - 2000 - The SIMBAD astronomical database. The CDS referenc.pdf:application/pdf},
 }
 
+@ARTICLE{wainer_catalog_2023,
+       author = {{Wainer}, Tobin M. and {Zasowski}, Gail and {Pepper}, Joshua and {Wagg}, Tom and {Hedges}, Christina L. and {Poovelil}, Vijith Jacob and {Fetherolf}, Tara and {Davenport}, James R.~A. and {Christodoulou}, P. Marios and {Dinsmore}, Jack T. and {Patel}, Avi and {Goold}, Kameron and {Gibson}, Benjamin J.},
+        title = "{Catalog of Integrated-light Star Cluster Light Curves in TESS}",
+      journal = {\aj},
+     keywords = {Star clusters, Time series analysis, Light curves, Variable stars, 1567, 1916, 918, 1761, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2023,
+        month = sep,
+       volume = {166},
+       number = {3},
+          eid = {106},
+        pages = {106},
+          doi = {10.3847/1538-3881/ace960},
+archivePrefix = {arXiv},
+       eprint = {2307.09510},
