### 2023-08-22 856494c
Update setup.cfg to match main
