+++ setup.cfg
@@ -18,7 +18,7 @@ long_description_content_type = text/markdown
 python_requires = >=3.9
 packages = find: 
 install_requires = 
-    numpy
+    numpy <1.24
     pip
     matplotlib
     astropy
@@ -29,7 +29,7 @@ install_requires =
     pandas
     tqdm
     h5py
-    lightkurve
+    lightkurve==2.4.0
     tables
     statsmodels
     imageio
