### 2023-08-22 #5605 @tobin-wainer
I have updated the JOSS paper with the citation of the AAS paper, which was recently [published](https://ui.adsabs.harvard.edu/abs/2023AJ....166..106W/abstract).
