+++ paper.bib
@@ -359,7 +359,7 @@ Type: article},
 	file = {Full Text PDF:/Users/Tobin/Zotero/storage/3NLKEDRJ/Wenger et al. - 2000 - The SIMBAD astronomical database. The CDS referenc.pdf:application/pdf},
 }
 
-@ARTICLE{wainer_catalog_2023,
+@article{wainer_catalog_2023,
        author = {{Wainer}, Tobin M. and {Zasowski}, Gail and {Pepper}, Joshua and {Wagg}, Tom and {Hedges}, Christina L. and {Poovelil}, Vijith Jacob and {Fetherolf}, Tara and {Davenport}, James R.~A. and {Christodoulou}, P. Marios and {Dinsmore}, Jack T. and {Patel}, Avi and {Goold}, Kameron and {Gibson}, Benjamin J.},
         title = "{Catalog of Integrated-light Star Cluster Light Curves in TESS}",
       journal = {\aj},
