+++ paper.bib
@@ -373,3 +373,4 @@ Type: article},
           doi = {10.3847/1538-3881/ace960},
 archivePrefix = {arXiv},
        eprint = {2307.09510},
+}
