### 2023-08-24 #5605 @ryanoelkers
> Hi @ryanoelkers, thanks for your comments.
>
> 1. This command is for formatting the figures in the jupyter notebook. We have added a comment to all the tutorial notebooks saying this command is not required, and only applicable to jupyter notebooks. We would prefer to keep it, as the figures look better on the docs website, but we can remove if necessary.
> 2. After reading this I re-ran the tutorial myself and also got 2 peaks. I realized that the tutorials were actually run on a previous version of elk, and with the latest version release this number changed. The tutorials are now all run with the latest version.
>
> While going through these comments we also added an upper limit on the numpy version since the latest release has breaking changes with the current lightkurve version.

Thanks @tobin-wainer ! Things looks good now.
