### 2023-08-24 #5605 @ryanoelkers
@tobin-wainer thanks for your hard work so far! A few other general comments I noticed.

1. Are you planning to review (and possibly accept) pull requests? If so, you may want to state that explicitly in the documentation.
2. In a similar vein, I couldn't find where folks could go to alert you of issues (either with bugs or when having trouble during installation). I realize the easiest method may just be to open an issue on GitHub, but it never hurts to say that explicitly in the documentation. Not all folks will know that "issues" exist on GitHub with being told about them :).
3. It looks like you still have a "TO DO" active in the documentation for citations. I realize some of that will occur once the JSOSS paper is published, but you may want to clean that up or add text without the various hyperlinks to references.
