### 2023-08-28 f4ae02d
fix cite page styling
