+++ docs/pages/cite.rst
@@ -2,28 +2,26 @@
 Citing elk
 **********
 
-If you use `elk` in a scientific publication, we ask for you to please cite. `elk` is jointly published in 'AJ <https://ui.adsabs.harvard.edu/abs/2023AJ....166..106W/abstract>'_ and JOSS (update link when accepted). <b>We ask you to please cite both papers.</b>
+If you use ``elk`` in a scientific publication, we ask for you to please cite. ``elk`` is jointly published in `AJ <https://ui.adsabs.harvard.edu/abs/2023AJ....166..106W/abstract>`_ and JOSS (update link when accepted). **We ask you to please cite both papers.**
 
 Below are ready to use BibTex citations for each paper. We recommend you cite the AJ paper in the main text of the paper, and the JOSS paper in the software section.
 
-#. 
+.. code-block:: bash
+
     @ARTICLE{Wainer_AJ,
-       author = {{Wainer}, Tobin M. and {Zasowski}, Gail and {Pepper}, Joshua and {Wagg}, Tom and {Hedges}, Christina L. and {Poovelil}, Vijith Jacob and {Fetherolf}, Tara and {Davenport}, James R.~A. and {Christodoulou}, P. Marios and {Dinsmore}, Jack T. and {Patel}, Avi and {Goold}, Kameron and {Gibson}, Benjamin J.},
-        title = "{Catalog of Integrated-light Star Cluster Light Curves in TESS}",
-      journal = {\aj},
-     keywords = {Star clusters, Time series analysis, Light curves, Variable stars, 1567, 1916, 918, 1761, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
-         year = 2023,
-        month = sep,
-       volume = {166},
-       number = {3},
-          eid = {106},
-        pages = {106},
-          doi = {10.3847/1538-3881/ace960},
-archivePrefix = {arXiv},
-       eprint = {2307.09510},
- primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023AJ....166..106W},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
+        author = {{Wainer}, Tobin M. and {Zasowski}, Gail and {Pepper}, Joshua and {Wagg}, Tom and {Hedges}, Christina L. and {Poovelil}, Vijith Jacob and {Fetherolf}, Tara and {Davenport}, James R.~A. and {Christodoulou}, P. Marios and {Dinsmore}, Jack T. and {Patel}, Avi and {Goold}, Kameron and {Gibson}, Benjamin J.},
+         title = "{Catalog of Integrated-light Star Cluster Light Curves in TESS}",
+       journal = {\aj},
+      keywords = {Star clusters, Time series analysis, Light curves, Variable stars, 1567, 1916, 918, 1761, Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
+          year = 2023,
+         month = sep,
+        volume = {166},
+        number = {3},
+           eid = {106},
+         pages = {106},
+           doi = {10.3847/1538-3881/ace960},
+        adsurl = {https://ui.adsabs.harvard.edu/abs/2023AJ....166..106W},
+       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
 
 ADD ONCE JOSS paper is accepted.
