+++ docs/pages/feedback.rst
@@ -1,0 +1,13 @@
+**************************
+Feedback and Contributions
+**************************
+
+How to contribute
+*****************
+Got an idea for an improvement for ``elk``? We're so glad to hear it! Please create a fork of `our repository <https://github.com/tobin-wainer/elk>`_ and start a `pull request <https://github.com/tobin-wainer/elk/pulls>`_. We'll review your changes and hopefully merge them into ``elk``!
+
+How to give feedback
+********************
+Found a bug? Do you think there's another documentation page that we need? We'd love to hear your thoughts. Please `create a GitHub issue <https://github.com/tobin-wainer/elk/issues/new/choose>`_ using one of our automatic templates and we'll work on helping you out with any issues!
+
+Otherwise if you like the package then we'd appreciate if you could star it on GitHub and tell your astro-friends!
