### 2023-08-28 0033211
simplify user guide
