+++ docs/modules/plot.rst
@@ -1,13 +1,8 @@
+:html_theme.sidebar_secondary.remove: true
+
 ************************
 Visualisation (``plot``)
 ************************
 
-Introduction
-============
-
-TODO add intro about module
-
-API
-===
-
 .. automodapi:: elk.plot
+    :no-heading:
