+++ docs/modules/lightcurve.rst
@@ -1,14 +1,9 @@
+:html_theme.sidebar_secondary.remove: true
+
 *****************************************
 Lightcurve investigation (``lightcurve``)
 *****************************************
 
-Introduction
-============
-
-TODO add intro about module
-
-API
-===
-
 .. automodapi:: elk.lightcurve
     :no-inheritance-diagram:
+    :no-heading:
