### 2023-08-28 #5605 @tobin-wainer
Hi @ryanoelkers, thanks for the comments.

1. and 2.  Thanks for pointing this out. We are planning on accepting pull requests and issues. We have created a page on the docs with instructions, and links on how to do this via github. We have also made templates for specific types of issues.

3. Yes, thank you for mentioning this. The AJ paper was recently accepted, and we have added the citation for this paper. We will add the JOSS citation when it is appropriate to do so.
