### 2023-09-05 tobin-wainer/elk#100 @danhey
Installation fails on M1
**Describe the bug**
Installation of elk fails because of the tables dependency incompatibility (https://github.com/PyTables/PyTables/issues/1038)

**Example for reproducing**
Please include a minimum working example that reproduces the error, e.g.
```
pip install astro-elk
```
or directly setting up from the provided environment.yml

**Expected behavior**
Installation should proceed without issue. Instead, the `tables` installation fails due to missing libraries:
```
ld: library not found for -llzo2
```

**Solution**
The solution is straightforward -- the main branch on the Tables repo should be installed from source.
```bash
pip3 install git+https://github.com/PyTables/PyTables
```
This should be mentioned in the documentation somewhere, until PyTables updates their pip release.

 - OS: macOS 13.5
 - Version: pip
 - Run from: terminal

For JOSS: https://github.com/openjournals/joss-reviews/issues/5605
