### 2023-09-06 #5605 @danhey
##### Intro Comments

Hi all, thanks for producing this nice bit of software, I can see it becoming widely used for light curve extraction. The code is well documented with a few nice examples. I've split up my comments below into the relevant sections.

##### Installation:

Installation was a little difficult for me, and took about an hour of troubleshooting to sort out (not the authors fault!). You can see above that I opened a PR regarding the problem, but prior to that the Python environment (3.8) I created was not supported. I think it would be useful to mention supported Python versions and a workaround for the above problem in the docs.

##### Getting started & Tutorials:

All work fine with no issue. It would be nice to have some of the text explain what's happening when the quality test fails. e.g., the first example fails 2 quality tests, how come? Should we be worried? Can we force the light curve to extract anyway?

The getting started tutorial has a typo in the link at the bottom of the page. There are quite a few typos in the docs in general (e.g., https://elk.readthedocs.io/en/latest/tutorials/diagnostic_gif.html has aperature, peack, identttify, etc).

The image for the diagnostic_gif is missing.

##### General paper comments:

The paper is nicely written and clearly explains the niche of integrated light curves. However, I think the paper is missing a short discussion of other light curve extraction software (e.g., eleanor). Indeed, looking at the source code shows that this method relies strongly on lightkurves _RegressionCorrector_ -- this should be mentioned more clearly in the text! Right now the text strongly implies that the `elk` is doing "singular value decomposition methods and principle component analysis" from scratch, but that is not the case.

Finally, it's not unusual for JOSS papers to be illustrated with a figure. I know there's a lot of nice figures in the AAS version of this submission, but it might be worthwhile having a simple plot of an extracted light curve. This is totally up to the authors.

Great work all, I'm looking forward to seeing more results come from this tool!
