+++ paper.bib
@@ -374,3 +374,21 @@ Type: article},
 archivePrefix = {arXiv},
        eprint = {2307.09510},
 }
+
+@ARTICLE{feinstein_eleanor_2019,
+       author = {{Feinstein}, Adina D. and {Montet}, Benjamin T. and {Foreman-Mackey}, Daniel and {Bedell}, Megan E. and {Saunders}, Nicholas and {Bean}, Jacob L. and {Christiansen}, Jessie L. and {Hedges}, Christina and {Luger}, Rodrigo and {Scolnic}, Daniel and {Cardoso}, Jos{\'e} Vin{\'\i}cius de Miranda},
+        title = "{eleanor: An Open-source Tool for Extracting Light Curves from the TESS Full-frame Images}",
+      journal = {\pasp},
+     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2019,
+        month = sep,
+       volume = {131},
+       number = {1003},
+        pages = {094502},
+          doi = {10.1088/1538-3873/ab291c},
+archivePrefix = {arXiv},
+       eprint = {1903.09152},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019PASP..131i4502F},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
