+++ paper.md
@@ -40,7 +40,7 @@ aas-journal: Astrophysical Journal
 
 The term `integrated' refers to the method in which we are extracting blended photometry from an aperture, a practice common in the literature. Our approach does this extraction at many time steps to generate light curves, and we evaluate how the integrated light changes through time. Each pixel light curve is determined independently, and then summed to create the integrated light curve.
 
-Our correction techniques utilizes the bona fide pipeline `lightkurve`, and are able to account for spatially-transient scattered light from the Earth and Moon (using singular value decomposition methods and principle component analysis), known TESS telescope systems (encoding in co-trending basis vectors) and long-period astrophysical variability (through the application of a basis-spline model). Each of these correction techniques from the `lightkurve` package are flexible within `elk` and can be tuned or turned off entirely. `elk` easily wraps the multiple components from `lightkurve` into a single, easy to use package for integrated light curves, a use case not common within existing extraction pipelines like `lighkurve` or `eleanor` [@feinstein_eleanor_2019].
+Our correction techniques utilize the bona fide pipeline `lightkurve`, and are able to account for spatially-transient scattered light from the Earth and Moon (using singular value decomposition methods and principle component analysis), known TESS telescope systems (encoding in co-trending basis vectors) and long-period astrophysical variability (through the application of a basis-spline model). Each of these correction techniques from the `lightkurve` package is flexible within `elk` and can be tuned or turned off entirely. `elk` wraps the multiple components from `lightkurve` into a single, easy to use package for integrated light curves, a use case not common within existing extraction pipelines like `lightkurve` or `eleanor` [@feinstein_eleanor_2019].
 
 `elk` additionally has a range of features to analyze integrated light curves. We include a wide variety of variability metrics from across astronomical subfields, including the Stetson J statistic and von Neumann ratio [@stetson_center_1994; @von_neumann_distribution_1941]. `elk` features simple and useful diagnostic capabilities, with flexible functions for the creation and visualisation of bootstrapped Lomb-Scargle periodograms and autocorrelation functions. Moreover, one can use `elk` to locate the precise sky position of a variable signature causing a feature in an integrated light curve, and identify likely candidate stars from SIMBAD [@wenger_simbad_2000]. Though we designed `elk` with the intention of studying cluster variability, it has the flexibility to be applied to *any* integrated study.
 
