### 2023-09-18 c6b3ed5
add sentence for figure reference.
