### 2023-09-18 bcbde1f
Trying a different sudo get for hdf5
