+++ .github/workflows/perform_tests.yml
@@ -40,7 +40,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        sudo apt-get install hdf5
+        sudo apt-get install libhdf5-serial-dev
         python -m pip install flake8 pytest coverage pytest-xdist nbmake pytest-cov
         if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
         python -m pip install .
