+++ docs/pages/install.rst
@@ -20,7 +20,10 @@ Installation
                 
                 pip install astro-elk
 
-        and you should be all set! Now it's time to learn how about `Getting Started <getting_started>`__ with ``elk``.
+        and you should be all set! Now it's time to learn how about `Getting Started <https://elk.readthedocs.io/en/latest/pages/getting_started.html>`__ with ``elk``.
+
+        .. note::
+            Please note, `elk` is supported for python 3.9+ environments. 
 
         .. note::
             If you also want to work with Jupyter notebooks then you'll also need to install jupyter/ipython to this environment!
@@ -61,7 +64,7 @@ Installation
 
             pip install .
 
-        and you should be all set! Now it's time to learn how about `Getting Started <getting_started.html>`__ with ``elk``.
+        and you should be all set! Now it's time to learn how about `Getting Started <https://elk.readthedocs.io/en/latest/pages/getting_started.html>`__ with ``elk``.
 
         .. note::
             If you also want to work with Jupyter notebooks then you'll also need to install jupyter/ipython to this environment!
