### 2023-09-25 #5605 @warrickball
Thanks @danhey and @ryanoelkers for your reviews! There are only editorial matters to worry about before the paper is published.

@tobin-wainer, three last things on the editorial side of the paper before we start wrapping this up.

First, there are a few typos. I did a quick spell check and found
* variabilty → variabil*i*ty (in the tags)
* astroseismology → ast*e*roseismology
You may want to check the article again for any more subtle slips, though none jumped out at me.

Second, many of the citations are malformed, e.g.,
* (e.g., Chaplin et al. (2013); Pinsonneault et al. (2018)).
* (e.g., Oelkers & Stassun (2018); Nardiello (2020); Higgins & Bell (2022)).
I realise this is probably because the citation mechanisms for Rmarkdown aren't familiar but I think they're documented [in the Pandoc docs](https://pandoc.org/MANUAL.html#citation-syntax), from which you should be able to get the desired effect. (Basically, remove the extra layers of brackets.) If you struggle, let me know and I'll try to devise the fixes, which I confess I don't know offhand myself.

Finally, the second paragraph starts with
> The term ‘integrated’ refers to the method in which we are extracting blended photometry from an aperture, a practice common in the literature.

To demonstrate it's a common practice, you should cite a few relevant examples of this practice at this point.
