+++ paper.bib
@@ -392,3 +392,39 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2019PASP..131i4502F},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{bastian_searching_2020,
+       author = {{Bastian}, Nate and {Lardo}, Carmela and {Usher}, Christopher and {Kamann}, Sebastian and {Larsen}, S{\o}ren S. and {Cabrera-Ziri}, Ivan and {Chantereau}, William and {Martocchia}, Silvia and {Salaris}, Maurizio and {Asa'd}, Randa and {Hilker}, Michael},
+        title = "{Searching for multiple populations in the integrated light of the young and extremely massive clusters in the merger remnant NGC 7252}",
+      journal = {\mnras},
+     keywords = {galaxies, star clusters, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2020,
+        month = may,
+       volume = {494},
+       number = {1},
+        pages = {332-337},
+          doi = {10.1093/mnras/staa716},
+archivePrefix = {arXiv},
+       eprint = {2003.03428},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.494..332B},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{krumholz_star_2019,
+       author = {{Krumholz}, Mark R. and {McKee}, Christopher F. and {Bland-Hawthorn}, Joss},
+        title = "{Star Clusters Across Cosmic Time}",
+      journal = {\araa},
+     keywords = {Astrophysics - Astrophysics of Galaxies, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2019,
+        month = aug,
+       volume = {57},
+        pages = {227-303},
+          doi = {10.1146/annurev-astro-091918-104430},
+archivePrefix = {arXiv},
+       eprint = {1812.01615},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2019ARA&A..57..227K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
