### 2023-09-26 994e70d
Fix Typos and add Citations
