+++ paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+aas-doi: 10.3847/1538-3881/ace960 
 aas-journal: Astrophysical Journal
 ---
 
