+++ paper.md
@@ -38,7 +38,7 @@ aas-journal: Astrophysical Journal
 
 `elk` (int**E**grated **L**ight **K**urve) is an open-source Python package designed for downloading and correcting *integrated* light curves from the Full Frame Images (FFI) of TESS data. We provide novel approaches for the analysis of heavily blended light curves. Our correction is in the form of aperture photometry, building upon the methodology of the Python package `lightkurve` [@lightkurve_collaboration_lightkurve_2018]. 
 
-The term `integrated' refers to the method in which we are extracting blended photometry from an aperture, a practice common in the literature [e.g, @krumholz_star_2019; @bastian_searching_2020]. Our approach does this extraction at many time steps to generate light curves, and we evaluate how the integrated light changes through time. Each pixel light curve is determined independently, and then summed to create the integrated light curve.
+The term `integrated' refers to the method in which we are extracting blended photometry from an aperture, a practice common in the literature [e.g, @bastian_searching_2020; @krumholz_star_2019]. Our approach does this extraction at many time steps to generate light curves, and we evaluate how the integrated light changes through time. Each pixel light curve is determined independently, and then summed to create the integrated light curve.
 
 Our correction techniques utilize the bona fide pipeline `lightkurve`, and are able to account for spatially-transient scattered light from the Earth and Moon (using singular value decomposition methods and principle component analysis), known TESS telescope systems (encoding in co-trending basis vectors) and long-period astrophysical variability (through the application of a basis-spline model). Each of these correction techniques from the `lightkurve` package is flexible within `elk` and can be tuned or turned off entirely. `elk` wraps the multiple components from `lightkurve` into a single, easy to use package for integrated light curves, a use case not common within existing extraction pipelines like `lightkurve` or `eleanor` [@feinstein_eleanor_2019]. Our correction techniques can be visualized in Figure 2 of  [@wainer_catalog_2023]. 
 
