### 2023-09-26 #5605 @tobin-wainer
Hi @warrickball,

Thank you for these comments. I have gone through and _believe_ I have gotten all of the typos. I have also fixed the formatting issue for all citations with "e.g.," and they are correct now. Finally I have added two citations for the sentence you refer to, one being a highly cited review article.

Thanks you for the comments, and if anything else is needed, please let me know.
