### 2023-09-27 #5605 @warrickball
Thanks for the sweep! I've noticed there are still a few extra brackets around citations, e.g.
> (TIC, Stassun et al. (2018))

> (e.g., Oelkers & Stassun (2018); Nardiello (2020); Higgins & Bell (2022)).

> elk is jointly published in (Wainer et al., 2023),

I'll have a look into these myself tomorrow morning to see if I can fix them via PR. I'll also brush up on the next steps so I can help wrap this up ASAP.
