### 2023-09-28 #5605 @warrickball
Not sure why the missing DOIs are flagged. I've followed them and they look correct.
