### 2023-09-28 #5605 @danielskatz
@warrickball - they are flagged because they're not in the bib entries as doi lines - the author should check them and add them if they are correct.
