### 2023-09-28 #5605 @warrickball
Regarding the DOI for Higgins & Bell, this whole reference should be updated to the published version [(Higgins & Bell 2023)](https://ui.adsabs.harvard.edu/abs/2023AJ....165..141H). I've opened a PR to fix the remaining brackets for citations and add the DOI for von Neumann (1941).
