### 2023-09-29 1a98efe
Create CITATION.cff
