### 2023-10-02 #5605 @tobin-wainer
Hi @warrickball,
Thank you for your work and the PR.

I have merged the PR, and released Version 1.0 to reflect the changes in the review process. I have created a DOI ( DOI: 10.5281/zenodo.8393595), and verified the author list and ORCIDs. I have also created a .cff file to comply with with DOI.

If there is anything else needed, please let me know.
