### 2023-10-02 #5605 @warrickball
Hi, @tobin-wainer, thanks! Note that the review checklist is mostly for me but I appreciate that you've effectively ticked some things off for me. Note, however, that

> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * Make sure that the license listed for the archive is the same as the software license.

The author list on the Zenodo archive should match the JOSS paper but is currently only you and Tom Wagg. You should be able to amend the Zenodo entry without having to mint a new one. Note also that Zenodo hasn't detected the MIT license, which you can also change in the Zenodo metadata. (I'm guessing this is a limitation of the GitHub integration.)

> I have also created a .cff file to comply with with DOI.

Thanks! The editorialbot actually now mints a CITATION.cff for the JOSS paper on acceptance, which you'll see and may wish to use.
