### 2023-10-02 #5605 @warrickball
@tobin-wainer, note also that my PR didn't fix the reference to [Higgins & Bell](https://ui.adsabs.harvard.edu/abs/2023AJ....165..141H), which should be update to the published version. I presume that'll fix the missing DOI but you should double check.
