+++ paper.bib
@@ -275,18 +275,22 @@ a significant ultraviolet excess.},
 }
 
 @techreport{higgins_localizing_2022,
-	title = {Localizing {Sources} of {Variability} in {Crowded} {TESS} {Photometry}},
-	url = {https://ui.adsabs.harvard.edu/abs/2022arXiv220406020H},
-	abstract = {The Transiting Exoplanet Survey Satellite (TESS) has an exceptionally large plate scale of 21"/px, causing most TESS light curves to record the blended light of multiple stars. This creates a danger of misattributing variability observed by TESS to the wrong source, which would invalidate any analysis. We develop a method that can localize the origin of variability on the sky to better than one fifth of a pixel. Given measured frequencies of observed variability (e.g., from periodogram analysis), we show that the corresponding best-fit sinusoid amplitudes to raw light curves extracted from each pixel are distributed the same as light from the variable source. The primary assumption of this method is that other nearby stars are not variable at the same frequencies. Essentially, we are using the high frequency resolution of TESS to overcome limitations from its low spatial resolution. We have implemented our method in an open source Python package, TESS Localize (github.com/Higgins00/TESS-Localize), that determines the location of a variable source on the sky given TESS pixel data and a set of observed frequencies of variability. Our method utilizes the TESS Pixel Response Function models, and we characterize systematics in the residuals of fitting these models to data. Given the ubiquity of source blending in TESS light curves, verifying the source of observed variability should be a standard step in TESS analyses.},
-	urldate = {2022-08-02},
-	author = {Higgins, Michael E. and Bell, Keaton J.},
-	month = apr,
-	year = {2022},
-	note = {Publication Title: arXiv e-prints
-ADS Bibcode: 2022arXiv220406020H
-Type: article},
-	keywords = {Astrophysics - Instrumentation and Methods for Astrophysics},
-	file = {Full Text PDF:/Users/Tobin/Zotero/storage/CBU36RKG/Higgins and Bell - 2022 - Localizing Sources of Variability in Crowded TESS .pdf:application/pdf},
+       author = {{Higgins}, Michael E. and {Bell}, Keaton J.},
+        title = "{Localizing Sources of Variability in Crowded TESS Photometry}",
+      journal = {\aj},
+     keywords = {Astronomical object identification, CCD photometry, Time series analysis, Variable stars, 87, 208, 1916, 1761, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2023,
+        month = apr,
+       volume = {165},
+       number = {4},
+          eid = {141},
+        pages = {141},
+          doi = {10.3847/1538-3881/acb20c},
+archivePrefix = {arXiv},
+       eprint = {2204.06020},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023AJ....165..141H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @article{hattori_unpopular_2022,
