### 2023-10-02 #5605 @tobin-wainer
Hi @warrickball,

Thanks for the instructions on the zendo things. This is my first time working with that program, so I was rather confused. I believe the authors and License are now accurate.

I also just updated the Higgins & Bell reference which now has a DOI.
