### 2023-10-03 c698d70
Update CITATION.cff
