### 2023-10-03 #5605 @dfm
Note that the paper doesn't seem to be rendering on the publication page just yet. This happens sometimes because of a caching issue, but it should resolve itself in the next few hours. We can check back in if it's not showing up by tomorrow.
